test 004: do not use '-y' argument
'-y' argument is not (yet) supported by the ssm, because very few tools
supports it. Lvm command should support '-y' but actually only pvremove
does. Also fsadm still handles '-y' incorrectly in the way that is quite
dangerous. So as a result of that, I have removed '-y' argument from ssm
for now.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>