btrfs: Add snapshot support
This commit adds snapshot support for the btrfs backend. We create a new
class to manage informations about the btrfs snapshots, note that btrfs
does not provide any way to distinguish between snapshot subvolume and
regular subvolume, so we are using the subvolume name format to
recognize the snapshot. Note that snapshot in btrfs actually _is_ just a
regular subvolume with some blocks shared with its original subvolume.

We also add a new method to the BtrfsVolume class called "snapshot"
which handle the creation of the snapshot.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>