lvm: Add snapshot support
This commit adds snapshot support for the lvm backend. We create a new
class to manage information about the lvm snapshots. We also add a new
method to the LvsInfo class called "snapshot" which handles the creation
of the snapshot. Note that we only allow to create a linear snapshot,
however this should change in the future.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>