backend lvm: Convert lv name into the dev_name
Convert lv name into the dev_name before we use it in remove, resize, or
snapshot. That's because we never know what will be passed in by the
backend and lvm requires the volumes to be referenced by the path, or
the name with volume group in it e.g. VolumeGroup/VolumeName, or
/dev/VolumeGroup/VolumeName

Signed-off-by: Lukas Czerner <lczerner@redhat.com>