lvm: Do not show dev_free and dev_used if pv is not in pool
If the pv is not in the pool we do not need the dev_free and dev_used
information, because it is just not used at all. Removing it will also
remove it from the device listing which is desirable, because printing
this information might be confusing for the user since we are hiding the
physical volume abstraction.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>