Use have backends to define default pool name
Currently there was only one default name for the pool. However this may
cause some problems and confusion, so it is better to let each backend
to define its own default pool name and use it when the name of the pool
is not specified by the user.

Also fix all tests accordingly so that it sets the environment variable
accordingly.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>