btrfs: Test all btrfs devices for mount point
Currently only the first btrfs device is considered to be "real device",
so when testing whether the file system is mounted, only this device
would be used. However if the file system is mounted with another device
we will fail to recognize mounted file system. Fix this by trying all
the devices in the file system.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>