Add possibility to temporarily mount devices
Currently there was no way to temporarily mount devices, however it is
sometimes needed. For example btrfs can not resize nor create subvolumes
when the file system is not mounted. Though we have to make sure that we
properly unmount the temporary location before we exit the script, so
exit clean up function is registered to unmount and remove create
temporary locations.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>