btrfs: Create new pool when adding to non-existing pool
If user is trying to add devices to the non-existing pool, btrfs should
comply with ssm logic and create a pool with defined devices in it. The
fact that the btrfs pool is actually the volume itself seems to be
implementation detail. There is simply no way how to create btrfs pool
without the "root" volume.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>