btrfs: Add new devices into the pool prior subvolume creation
Currently adding a device into the pool before creating a volume or
subvolume was forbiddent for btrfs. But now there is no reason to forbid
this anymore so allow adding new device before attempt to create a new
subvolume. Although we still have to disallow creation the pool if the
pool does not exist, because we're going to create the pool in create()
routine.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>