Add new get_device method to checking for block device
Currently when adding devices into the existing pool, we only check
whether provided device is rally a block device. However the problem
might be when the block device name (path) is different from the one
we're using to track the block device internally.

Fix this by introducing the new get_bdevice() method which is using
_find_device_record() to actually find the device name we're using
internally.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>