misc: use /proc/self/mountinfo if exists
Add new function get_mountinfo() to get the information about mounts in
the system. However older systems might not have this file, so we will
use this only when it really exists and fall back to the /proc/mounts
otherwise. This also required some changes how the function is used.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>