btrfs: Show real mount point of the subvolume
Currently when listing btrfs subvolumes wrong mount point might be
shown. It is because we alway assumed that the subvolume mount point
is the same as the root subvolume mount point, but it might not be the
case since one can mount subvolumes to the other part of the file system
hierarchy.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>