doc: Add sphinx documentation directory
This commit adds documentation directory in doc/. Documentation is
generated using sphinx, which help us to avoid duplication of texts for
different type of documents (html, man, readme, ...). Documentation
is separated into small sections which are stored in the *.rst and *.txt
files which are then combined into different types of documents.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>