Add ssm.8 manual page
This commit adds manual page ssm.8 to the project. Manual page is
generated with sphinx from the doc/ directory. However since we do not
want to require sphinx as dependency for the installation of System
Storage Manager, it is better to add generated manual page to the source
code as well.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>