Add INSTALL file
This commit adds INSTALL file into the source files. It is generated by
sphinx from the doc/ directory and it contains information about how to
install System Storage Manager on the system. Even though README
contains this information as well, I think it is important enough to be
in separate file.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>