Add problem.py module for dealing with problems
This commit adds new interface (problem.py) for dealing with problem,
warnings and user input in generic way. We have defined a set of
problems, optionally with predefined questions for user, which can be
used from within the ssm to deal with problems.

For example, when the ssm hits notices that the volume it is trying to
remove is mounted it can just use this interface, where we automatically
detect whether to ask user about unmounting the volume, what to ask
exactly and return the result of user answer back to the ssm, or throw
an error. Problems are configurable and each backend can use the
ProblemSet() class to define its own set of problems without the need to
reimplement the whole infrastructure.

Note that this commit only adds this new interface but it is not used
anywhere in the ssm yet.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>