main: Wire in problem interface
Wire in the problem interface into the ssm. We initialize and pass the
options to the problem interface and make a use of the interface for
generic problems, errors and information.

This commit also reorganizes create and resize in order to make a use of
problem interface to ask user questions.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>