btrfs: Compute pool_free as difference between pool_size and fs_used
It is a bit tricky to set pool_free right. Because this value is used
when creating new subvolumes, but btrfs subvolumes does not actually
take any space to create (almost), so pool_free should rather. So
pool_free should rather refer to the difference between pool_size and
fs_used.

When the quotas will be used for defining btrfs subvolume size, then it
has to be changed.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>