Use Options() object to pass options between classes
Currently we are passing several options separately, but in the future
the amount of options can grow, so it is better to pass options a
class with variable.

This commit creates new class Options() with default options setting and
pass this options argument instead of separate arguments for each option.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>