Change how create and resize handles optional devices
Currently on resize we would add only the number of devices to cover the
size. However the volume might be in RAID setup and then it might fail
when not providing enough devices to resize it correctly.

Change this behaviour to always add all devices provided to the pool
before attempt to create or resize.

Also if the device can provided can not be added to the pool for some
reason, we should fail if used did not specified the size of the volume
to be created. The reason is that, that is we would used only devices
which are actually available, the result would be different from user
expectations. So the better options is to fail and let used deal with it
manually.

On the other hand, if the size is specified and the amount of usable
devices is sufficient to cover this size, then it is reasonable to
create the volume with usable devices, ignoring the rest. In case of
RAID setup, when the user have specified number of stripes, this needs
to be dealt with in backend.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>