Fix check for the root user
The check for the root user is wrong. Specifically the condition which
checks for the harmful arguments is wrong causing the possibility to run
commands that requires root permissions to be run without root
permissions.

Note that this is not a security problem! The command will simply fail,
but the check is there so that the command would not fail, but rather
inform user that he needs root permissions to run the script.

This commit fix the condition.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>