Allow ssm to be run locally without installing
Currently there is not way of running ssm without actually installing
it to the system, or changing python path to search for the modules in
the local directory.

This commit adds script bin/ssm.local which can be run locally with the
local version of the modules by changing PYTHONPATH and running the
bin/ssm script.

This commit also changes how bash tests are invoked so that it uses the
local version of the tool and ssm modules which is actually desired way
how test should work.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>