Remove itertools from dependencies
With this commit we remove itertools python module from the ssm
dependencies as it might not be present in some distributions and it is
easy enough to to include the couple of functions we're using directly
into misc.py.

Include functions chain(), izip(), compress() and permutations() from
the itertools python module into ssmlib/misc.py.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>