Catch RuntimeError when initializing backend
Currently then RuntimeError happens when initializing backend because
the command failed to complete properly, we just exit. But it would
actually make sense to skip such backend entirely (or its part) and
continue.

This commit catches RuntimeError exception when initializing backend to
be able to print out warning and continue normally.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>