Set LC_NUMERIC instead of LANG variable
If any of the tools we're using in ssm generates numerical output with
the decimal mark we need that mark to be a point not a comma. However
setting LANG environment variable is not enough since it will only be
used if LC_ALL is not set, or if any of the LC_* is missing.

This patch sets LC_NUMERICAL to C so we always get numerical point in
floating point numbers generates by the tools used in ssm.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>