Wrap table lines when terminal width is too small
When the terminal window is not wide enough for the whole table
we can try to fix the formatting by wrapping the line right after
the first attribute. On the other hand, if it would not help anyway
then can skip the line wrapping since it would only make things
worse.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>