main: Fix resizing device without pool
It is possible to resize file systems on regular devices, or volumes
which backed does not support pooling. However currently resize of such
device would fail because we would try to use pool which does not exist.

Fix this by checking for non existing pool and supply default values if
needed.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>