btrfs: Rework how subvolume info is gathered
Currently there is a problem with subvolume parsing in the case that the
subvolume is mounted somewhere outside root subvolume mountpoint, or in
the case that the root subvolume is not mounted at all. In those cases
subvolume mount points might not be right or the subvolume might not be
even found properly.

This commit reworks the way how we parse the subvolume information. It
has been based on pathes for Tomas Marek.

Also fix unittests to match the reality.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>