btrfs: Use regex to gather info about subvolumes
Currently we rely on static positioning of the output of 'btrfs
subvolume list' however the outptu format can change (and have indeed
changed) by adding more information into the output. To avoid problems
search for the information we need with regular expression.

With this we're little bit more resistant to the btrfs subvolume list
output.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>