btrfs: Notify udev when device is removed from fs
Some versions of btrfs does not notify udev about device removal and
with newer version of liblkid it might happen that the by-uuid and
by-label links still point to the device which is no longer part of the
file system effectively preventing it to mount.

Fix this by sending change event to the device.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>