btrfs: Check path when creating subvolume
Currently user can try to create subvolume of the name 'hello/world'
so we attempt to create such subvolume under root subvolume, however in
the case that hello/ does not exist, then 'subvolume create' command
would fail.

We should check whether the path exists before attempt to create the
subvolume and error out in cat that the path does not exist. Add
appropriate test for this case into bashtests.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>