btrfs: Fix btrfs subvolume removal code
If we try to remove subvolume which is not specifically mounted, we
should mount the root subvolume and then attempt to remove it. Also if
we try to remove subvolume which has been explicitly mounted we should
ask user to umount it firs and then remove it.

This commit also adds appropriate tests for this case.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>