ssmlib: Add new mount command
This commit adds new 'mount' command which is intended to mount the
volume on specified directory in the same way as mount(8) does. This
makes user life easier, because volume names does not always correspond
to the mountable "strings" and ssm can do the translating for this user
since it has all the information.

For example, you can have btrfs file system with several subvolumes.
Mounting the file system (btrfs_pool) or the subvolume
(btrfs_pool:vol001) would require user to find out the device and
subvolume id in case of subvolume. Now one can simply type

	ssm mount btrfs_pool /mnt/test

to mount the while btrs file system, or

	ssm mount btrfs_pool:vol001 /mnt/test

to mount the subvolume. One can also specify mount options in the same
way as with mount(8).

Signed-off-by: Lukas Czerner <lczerner@redhat.com>