resize: Do not free device from the pool if not necessary
Currently if the user provided devices to use when resizing a volume and
some of the devices provided are already used in the different pool we
would attempt to remove that device from the pool even in the case that
the resize will eventually fail because we're resizing to the same size.

Since we have this information before _filter_device_list() we should
check for the new and old size before we attempt to work with provided
devices.

RHBZ 987228
Signed-off-by: Lukas Czerner <lczerner@redhat.com>