tests: Fix failure of list_table in some tests
Currently is environment variable LINES is not set in the system we
would use ioctl to get a terminal size. This means that COLUMNS size set
in the lib/test does not have any effect. As a result of this some tests
doing check list_table might fail.

Fix this by exporting LINES variable as well. It's not really used so
the value does not matter, however we need for COLUMNS to take effect.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>