ssmlib: Fail if there is no enought space to create volume
Currently in some cases when creating volume we might end up trying to
create volume even though we already know there is not enough space in
the pool or in the provided devices, so this patch fixes that.

Also when the devices are provided, we should really only count the size
of those devices and not the entire pool size, because it might lead to
unexpected failures. So if user provides devices and the size, the size
of the volume must fit the provided devices.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>