btrfs: Always show btrfs snapshot properly
up until now we relied on the actual subvolume name to recognize
snapshot subvolume. This was due to the fact that btrfs was not able to
give us this information on whether the subvolume is snapshot subvolume
or regular subvolume.

With newer version of btrfs-progs we're able to list snapshot subvolumes
so this commit takes advantage of this.

However btrfs-progs can not list only subvolumes which are not in fact
snapshot, so we have to filter it out manually in ssm so that one
subvolume does not appear in both places volume list and snapshot list.

RHBZ 986125
Signed-off-by: Lukas Czerner <lczerner@redhat.com>