bashtests/014-lvm-resize.sh: add 014-lvm-resize.sh
This case is intended to replace 004-lvm-resize.sh
The old 004-lvm-resize.sh script has several problems:
The logic of the test script is not very clear and not very systematic.
Many tests don't check size after resize
Some filesystem related tests are only done on one specific filesystem
Since the noninteracive mode is set by the script, tests with echo 'y' | ...
shouldn't work
Ext2 filesystem is never tested

So this test case is intended to solve these problems
The tests are mainly executed by _test_resize like the old script. This function
can be execute with or without a filesystem, so if you want to change the
procedure, just change the funtion would be enough
We seperate the test function in tree scenarios:
1. Resize with no device
2. Resize with device from no pool
3. Resize with device from other pool

Each scenario, we test to increase, maintain and decrease size of volume. We use
all possible methods to change the size.

After the function, we try to test it with no filesystem and different kinds of
filesystem.

At the end, there are two minor invalid argument check in addition to the
original test

[ Fix some problems with the test, update it and merge it into
004-lvm-resize test --Lukas]

Signed-off-by: Jimmy Pan <jipan@redhat.com>