ssmlib: Allow to check volume through backend
Allow to use check command not only for file system check, but also to
check volume itself using backend specific code. The firs user of this
is btrfs. Note that even though btrfs is a file system we do not treat
it as one, for ssm this is just another backend.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>