misc: fix get_signature not to return error messages
Currently get_signature() will return whatever comes out of the blkid
includin possible error messages. This could be confusing for the caller
or even wrong in some cases because there is not way to tell whether
it's a valid signature or not.

Fix this by returning None if blkid exited with error.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>