ssmlib: Fix parsing size in create and shapshot command
Currently it is possible to define size with -+ signs even when creating
new volume or snapshot. Also when the sice argument is not a valid size
just a plain Exception is raised. Fix this by moving the check into
argument parser and disallow +- signs in the size and return appropriate
error.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>