ssm: force btrfs device add
Btrfs is broken. It does not recognize the btrfs signature properly and
even when the device has been wiped with wipefs btrfs would still claim
that the device actually is 'btrfs' device which it is not!

We've already fixed the case for mkfs.btrfs, howeve btrfs recently added
the check for 'btrfs device add' as well, so we need to do the same
thing. So make _can_btrfs_force() more generic and use it in extend()
method as well.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>