misc: Fix get_device_size to work with partitions correctly
Currently get_device_size() searches for the size in sysfs specifically
(/sys/block/<devname>/size) however that is not the right palce for
partitions since partitions are subdirectories in /sys/block/<devname>
directory.

So change get_device_size() to get major and minor numbers of the device
in question and use sysfs file /sys/dev/block/<major>:<minor>/size to
get the size.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>