misc: Use wipefs -a in misc.wipefs() helper
Currently we only wipe the first signature we can find with wipefs.
However wipefs has an ability to wipe all the signatures of given type
from the device (yes this is once again about btrfs) with option '-a'.
So use that instead. This also simplify the wipefs helper a lot.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>