lvm: Ask user when shrining inactive logical volume
When resizing inactive logical volume we can not tell what is present on
the volume itself. If there is a file system we will not get notify and
we could possibly lose data if an error has been made.

Fix this by checking whether the volume is active on the resize
(srhinking specifically), warn user about this fact and let him abort
the action.

Also add appropriate unittests and bashtests.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>