lvm: don't display snapshot size for thin snapshots
lvm thin snapshots don't report a "snap_percent" field when lvs is
executed. As a result of this, their "snap_size" in ssm is empty.
Previously, running "ssm list" with lvm thin snapshots present would
cause the program to crash while trying to cast "" as a float. This
patch prevents the crash by not trying ot calculate the snapshot
size for thin snapshot volumes.

Signed-off-by: Ricky Ayoub <rayoub@redhat.com>