Fix traceback when calling 'ssm list' with empty dm tables
If there is a device with an empty table in the output of "dmsetup table", commands like "ssm list", "ssm list vol" and "ssm list filesystems" will produce the following backtrace:

Traceback (most recent call last):
  File "/usr/bin/ssm", line 48, in <module>
    sys.exit(main.main())
  File "/usr/lib/python2.7/site-packages/ssmlib/main.py", line 1855, in main
    args.func(args)
  File "/usr/lib/python2.7/site-packages/ssmlib/main.py", line 1215, in list
    self.vol.ptable(more_data=self.dev.filesystems())
  File "/usr/lib/python2.7/site-packages/ssmlib/main.py", line 840, in vol
    self._volumes = Volumes(options=self.options)
  File "/usr/lib/python2.7/site-packages/ssmlib/main.py", line 736, in __init__
    self._data['crypt'] = crypt.DmCryptVolume(options=self.options)
  File "/usr/lib/python2.7/site-packages/ssmlib/backends/crypt.py", line 167, in __init__
    dm['type'] = array[3]
IndexError: list index out of range

In this situation of an empty table, the array where we store each line of the "dmsetup table" output will contain only one element - the name of the device - and therefore we will get an IndexError when we try to access the inexistent fourth element in the array, which corresponds to the map type.
The proposed patch checks the size of the array and, when is 1, prints out a warning message and continues to the next entry.

Signed-off-by: Jose Castillo <jcastillo@redhat.com>
Signed-off-by: Lukas Czerner <lczerner@redhat.com>