String and unicode related changes to support Python 3
Function 'print something' is removed in Python 3. As a result, 'print(something)'
is used instead for Python from now on and it's backward compatible for Python 2.

By default, 'print()' function will add the '\n' at the end of any strings. So, if
the end line is not desirable, sys.stdout.write() is used instead as the option to
add the comma (,) is not supported in Python 3. Likewise, sys.stderr.write() is used
for printing error message as 'print >> sys.stderr' is no longer supported. However,
'sys.stdout/stderr.flush()' is needed to make sure the outputs are printed.

The 'str()' function is not working correctly in Python 3 if the input is
byte value as it will simply print the string with '\b' prefix. As a result,
a new function '__str__()' is added to use the UTF-8 encoding in Python 3.

The 'u' prefix is no longer supported in Python 3 (though it is reintroduced in
Python 3.3). Therefore, the 'u()' is added to replace the 'u' prefix.

Signed-off-by: Vu Dinh <vdinh@redhat.com>
Signed-off-by: Lukas Czerner <lczerner@redhat.com>