Exception handling changes to support Python 3
Exception syntax is changed in Python 3 as the ',' (comma) will cause syntax
error. The syntax 'as' is used instead and it's backward compatible in Python 2.

The argparse behavior is changed a bit in Python 3 as missing subcommands may get past
without getting caught for "too few arguments" like in Python 2. The 'dest' error message
is added and make sure the subcommands are required.

Signed-off-by: Vu Dinh <vdinh@redhat.com>
Signed-off-by: Lukas Czerner <lczerner@redhat.com>