Dictionary method changes to support Python 3
In Python 3, 'itervalues()', 'iterkeys()' and 'iteritems()' are all removed.
As a result, to support Python 3, either remove those functions or use the
substitute functions such as 'values()', 'items()', 'values()'. However,
in Python 3, those functions return a new view of dictionary instead of
a list. So, in case the list is desired, the list cast 'list()' is used.

Signed-off-by: Vu Dinh <vdinh@redhat.com>
Signed-off-by: Lukas Czerner <lczerner@redhat.com>