Map, range, next functions and iterator changes to support Python 3
In Python 3, both 'map()' and 'range()' functions return an iterator instead of
a list. For that reason, if the list is desired, the list cast 'list()' is used
to get a list instead of an iterator.

The function 'iter.next()' is obsolete in Python 3 and is replaced by a built-in
function 'next()' (standalone). As a result, the new function '__nexy__()' is
added to support both Python 2 and 3.

The function 'izip()' from 'itertools' library is changed in Python 3. So, the
new 'izip()' from Python 3 'itertools' is replaced the old one if the Python 3
interpretor is deteted.

Signed-off-by: Vu Dinh <vdinh@redhat.com>
Signed-off-by: Lukas Czerner <lczerner@redhat.com>