ssm: Create a mount directory if it doesn't exist
Currently if the directory we want to mount to does not exist (using
create or mount command) it will fail immediately. This patch adds the
ability to create said directory. If the directory to mount to does not
exist ssm will ask user whether to create the directory or not. If
forced, or in non interactive mode, the directory will be created
automatically.

Moreover add some tests for this functionality.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>