ssm: add a workaround for lvm and exported volumes
When lvm lvs command is run on exported volumes, lvm ends with return
code 5 instead of 0 on a clean run. Because LVM people said that
this is intentional behaviour, it needs to be worked around in SSM.

This patch adds a condition that detects the mentioned issue and sets a
can_fail flag, so the error code is ignored.

Signed-off-by: Jan Tulak <jtulak@redhat.com>