ssmlib: change how stdin data are passed to subprocesses
This is fixing an issue where passing a stdin data to a subprocess (e.g.
cryptsetup) led to (probably) a race condition, and ssm would crash with
stdin being closed before any data were written. Simplify the call and
use a more direct approach which does not seem to suffer the race
condition.

Signed-off-by: Jan Tulak <jtulak@redhat.com>