crypt: change how passphrase is acquired
A solution to bug https://bugzilla.redhat.com/show_bug.cgi?id=1141871

SSM created a LV and then tried to create encrypted volume on top of
that. However, if the passphrase was weak, cryptsetup failed and the
user was left with half of the changes done and had to undo them
manually.

Fix this issue by asking for the password in advance by SSM itself
(instead of letting cryptsetup to ask on its own) and attempting to
encrypt a small temporary file in tmp with cryptsetup. That way, we can
be sure what cryptsetup thinks about the passphrase before we do any
operation on real devices.

Signed-off-by: Jan Tulak <jtulak@redhat.com>