ssm: split table printing into a standalone function
Split psummary (former ptable) into two functions. One that remains
a method of Storage class and prepares the data to be printed, and then
a new misc.ptable, which does the actual printing, but can print any data.

This is paving a way for a new Storage.pinfo method, which needs a
similar printing capabilities, but will differ in the data it prints.

This commit adds also tests for the new ptable function.

Signed-off-by: Jan Tulak <jtulak@redhat.com>