ssm: add migrate command
Add a new command to ease migration of data from a device to a device.

For general devices, dd is used. This command is more useful for lvm and
btrfs pools, where it uses their specific tools to migrate data between
physical volumes. Note that it is 'migrate' - the data can be removed
from the original device by these tools after a successful transfer.

Signed-off-by: Jan Tulak <jtulak@redhat.com>