ssm: add bash-completion script
From: Ondra Benes <obenes0@centrum.cz>

This completion script supports ssm operations and utilizes ssm itself
to suggest possible devices, not only commands. For example, 'ssm remove
<tab>' will use ssm list to offer devices that can be removed.

This advanced functionality works only for root user. If ssm is used
under non-root user with sudo, only commands are suggested (as we can't
run ssm without root).

Right now, this script is not automatically installed by the python
install scripts - it is up to distribution packagers to put it to the
correct place when building a package.

Resolves #4
Signed-off-by: Ondra Benes <obenes0@centrum.cz>
Signed-off-by: Jan Tulak <jtulak@redhat.com>

---
v3:
- hardcoded paths replaced with $(which xxx)