mount: Check the volume argument in parser
Currently there are no checks on the volume argument of mount command.
It means that we'd call mount with whatever, even attempt to create the
mount directory even if the volume itself does not exist.

Check that the volume is either the ssm volume or at least a block
device, otherwise print error and exit.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Reviewed-by: Jan Tulak <jtulak@redhat.com>