btrfs: send udev change event after some more btrfs operations
When working with btrfs there often seems to be a problem with udev
interfering right after some operation on the file system (migrate,
create, remove, reduce extend). I suspect that this happens mostly when
the operation is performed in kernel, however btrfs does not explicitly
send a udev change event. This needs to be fixed in kernel, meantime we
need a workaround (especially for migrate)

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Reviewed-by: Jan Tulak <jtulak@redhat.com>