lvm: Remove pool_name key if pv not in pool
When pv is not in volume group the pool_name will still be present in
dictionary and it's value is going to be empty. Howeve this might
be confusing because in several places we check for existing pool simply
by checking for existence of 'pool_name' key. So remove it.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Reviewed-by: Jan Tulak <jtulak@redhat.com>