crypt: crypt backend need to have a pool in database
Usually pool is the main handle for the backend, however crypt backned
has no pool, even though the volumes do reference pool_name.

So far we did not need it, however with some operations such as migrate
we need to recognize that the device belongs to a valid pool and call
into the backend to do some work. For that we need to have a valid pool.
Create it.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Reviewed-by: Jan Tulak <jtulak@redhat.com>