ssm: Rework migrate command
Migrate command was not working as expected in many ways so here is a
rework of it. The 018 bashtest is also reworked and significantly
extended to cover more possible scenarios and test the actual results of
the operation.

Lvm migrate now does not remove the pv from the volume group as I feel
there are perfectly valid scenarios where user would like to migrate
used pv to a different device and repurpose the source device.

unittests for lvm and btrfs are only changed so that they not fail,
however the unittests needs to be extended for btrfs and lvm as well.
Generic migrate unittests (test_ssm.py) are missing entirely.

Also some more fine grained testing on plain devices is completely
missing at the moment because in this way the testing filter actually
interfere. Those tests are commented out in the 018.So there is still
some work to be done.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Reviewed-by: Jan Tulak <jtulak@redhat.com>