lvm: index thinpools as vg-lv pair instead of just lv name
This fixes an issue where ssm incorrectly displayed only one of
the same-name thinpools in different VGs, as the thinpool name was used
as index key. This issue is fixed by using "vg/thinpool" as the index
instead of just the "thinpool" name.

Eg. in vg1/thinpool and vg2/thinpool, ssm saw only vg1/thinpool.

Signed-off-by: Jan Tulak <jtulak@redhat.com>

---
CHANGES

v2:
- Simply changing the index key in lvm led to inability to correctly
detect the thin pool for manipulation with it later on, which was caught
by the existing tests. So, change the few other places where the index is
used to look up for the new compound form instead of the old name.