ssmlib: missing pwquality should not crash the whole ssm
The pwquality python binding is needed only for creating encrypted
volumes, and not every distribution has it easily available. Simplify
the life of the users by requiring it only for the creation of encrypted
volumes, and even then allow the user to force the password if pwquality
is not available.

Closes #5
Signed-off-by: Jan Tulak <jtulak@redhat.com>

---
CHANGES:
v2
* instead of disabling all creations, ask the user if ssm should force
the password