ssmlib: test for identity and not just a name match
When building a graph of items for the info command
(parents/children/aliases/...), use an identity test to detect
duplicate items, rather than just a name.

The issue with name is that two items can share the name, but represent
a different object or a type of object.

This and the two previous patches together fixes
RHBZ 1613658
RHBZ 1649875
Closes #6

Signed-off-by: Jan Tulak <jtulak@redhat.com>