tests: fail test.py on unittests failure
If unittests ends up with a failure, exit with an error code. There is
no point in running a ton of bash tests when something is apparently
wrong.

Until this patch, we printed the errors out, but did not set the exit
code and continued, so if bash tests were not run, the exit code was 0.

Signed-off-by: Jan Tulak <jtulak@redhat.com>