bashtests: add --units h to lvs/pvs/vgs
Use --units h to avoid '<' rounding indicator that can appear in the
default unit.

From man pvs:
--units r|R|h|H|b|B|s|S|k|K|m|M|g|G|t|T|p|P|e|E
   All sizes are output in these units: human-(r)eadable with '<'
   rounding indicator, (h)uman-readable, (b)ytes, (s)ectors, [snip]

The values should not change, but if there is some small
platform-specific difference that needs rounding, it won't break the tests.

Signed-off-by: Jan Tulak <jtulak@redhat.com>

---

Lukas, do you think that this is a good approach? Or should we change it
to --units k and demand exact sizes?

Thanks,
Jan