ssmlib: verify requirements before committing to any action
Check if all used backends have their requirements before making any
changes to storage. This prevents situations where ssm e.g. aborts with
missing cryptsetup after it already created LVM pool and volumes.

The verification is added to operations that can require more than one
backend and we want to be sure that they do not fail in the middle. For
other operations like check, there is no penalty in discovering the
missing dependency on the fly, so we keep them simpler.

Signed-off-by: Jan Tulak <jtulak@redhat.com>